add the Tmux configuration
This configuration includes the appearance of the program (the theme is
Catppuccin Mocha) and some keyboard shortcuts. The status bar is hidden
by default, but, it can be shown with CTRL-B + Z.

Also included is a script to launch the program in the last session if
it is not attached, otherwise it will create a new session, very useful
in my launcher.