add the Neovim configuration
This includes the expected Catppuccin theme (in its Mocha flavor) and
some configurations using the Lua language.

Lazy is used as plugins manager, and among the plugins to highlight are
`vim-go`, `Ultisnips` and `lualine`.

This configuration is not yet complete, but, it already makes the editor
more comfortable to use.